requirements.txt
streamlitgoogle-generativeairequests